+++ lib/pub_sub_model_sync/base.rb
@@ -13,5 +13,12 @@ module PubSubModelSync
         config.log message, kind
       end
     end
+
+    def retry_error(error_klass, qty: 2, &block)
+      @retries ||= 0
+      block.call
+    rescue error_klass => _e
+      (@retries += 1) <= qty ? retry : raise
+    end
   end
 end