+++ lib/pub_sub_model_sync/message_processor.rb
@@ -2,7 +2,7 @@
 
 module PubSubModelSync
   class MessageProcessor < PubSubModelSync::Base
-    attr_accessor :payload
+    attr_accessor :payload, :raise_error
 
     # @param payload (Payload): payload to be delivered
     # @Deprecated: def initialize(data, klass, action)
@@ -24,23 +24,25 @@ module PubSubModelSync
     def run_subscriber(subscriber)
       return unless processable?(subscriber)
 
-      subscriber.process!(payload)
-      res = config.on_success_processing.call(payload, subscriber)
-      log "processed message with: #{payload.inspect}" if res != :skip_log
+      retry_error(ActiveRecord::ConnectionTimeoutError, qty: 2) do
+        subscriber.process!(payload)
+        res = config.on_success_processing.call(payload, { subscriber: subscriber })
+        log "processed message with: #{payload.inspect}" if res != :skip_log
+      end
     rescue => e
-      print_subscriber_error(e)
+      raise_error ? raise : print_subscriber_error(e, subscriber)
     end
 
     def processable?(subscriber)
-      cancel = config.on_before_processing.call(payload, subscriber) == :cancel
+      cancel = config.on_before_processing.call(payload, { subscriber: subscriber }) == :cancel
       log("process message cancelled: #{payload}") if cancel && config.debug
       !cancel
     end
 
     # @param error (Error)
-    def print_subscriber_error(error)
+    def print_subscriber_error(error, subscriber)
       info = [payload, error.message, error.backtrace]
-      res = config.on_error_processing.call(error, payload)
+      res = config.on_error_processing.call(error, { payload: payload, subscriber: subscriber })
       log("Error processing message: #{info}", :error) if res != :skip_log
     end